scp-294 6

+++ src/scp-294/common/database.go
@@ -8,34 +8,34 @@ import (
 
 var Db *sql.DB
 
-func Connect() {
+func Connect() error {
 	var err error
 	connStr := "./scp294.db"
 	Db, err = sql.Open("sqlite3", connStr)
 	if err != nil {
-		logger.Log(err.Error())
-		return
+		return err
 	}
 
 	ctx := context.Background()
 	err = Db.PingContext(ctx)
 	if err != nil {
-		logger.Log(err.Error())
-		return
+		return err
 	}
 	logger.Log("Database Connected!")
+	return nil
 }
 
-func InitDatabases() {
-	Db.Exec(`DROP TABLE IF EXISTS t_group`)
-	Db.Exec(`DROP TABLE IF EXISTS record`)
+func InitDatabases(forceClean bool) {
+	if forceClean {
+		Db.Exec(`DROP TABLE IF EXISTS t_group`)
+		Db.Exec(`DROP TABLE IF EXISTS record`)
 
-	Db.Exec(`CREATE TABLE IF NOT EXISTS t_group(
+		Db.Exec(`CREATE TABLE IF NOT EXISTS t_group(
 			Id INTEGER PRIMARY KEY AUTOINCREMENT,
 			Name VARCHAR(32) NOT NULL
 		)`)
-	Db.Exec(`INSERT INTO t_group (Id, Name) values (0, 'Default');`)
-	Db.Exec(`CREATE TABLE IF NOT EXISTS record(
+		Db.Exec(`INSERT INTO t_group (Id, Name) values (0, 'Default');`)
+		Db.Exec(`CREATE TABLE IF NOT EXISTS record(
 			Id INTEGER PRIMARY KEY AUTOINCREMENT,
 			Name VARCHAR(32) NOT NULL,
 			ConvertType VARCHAR(32) NOT NULL,
@@ -43,5 +43,6 @@ func InitDatabases() {
 			OutputData TEXT, 
 			GroupId INTEGER
 		)`)
-	Db.Exec(`CREATE UNIQUE INDEX recordIndex ON record (GroupId, Name);`)
+		Db.Exec(`CREATE UNIQUE INDEX recordIndex ON record (GroupId, Name);`)
+	}
 }
